HBSD: fix bhyveload memory allocation error
Details:

We changed the VM_MIN_ADDRESS from 0 to 64k to prevent mapping bellow the
64k address from user-space, this change prevents NULL pointer derefs.
In vm_* functions checking for MAX(sv_minuser, PAGE_SIZE) in our case
this always evaluated to 64k, but this check requiered only for
user-space programs.
In bhyve case, in the vmm module created new vmspace, but this must to
be beginning from address 0 to properly emulate things. In the original
implentation they used VM_MIN_ADDRESS to set this addres, but we
(HardenedBSD) changed this constant, to resolve this, change the
VMSPACE_ALLOC() parameter from VM_MIN_ADDRESS to 0.

Tested on Core i5-4200U CPU with FreeBSD installer.

Signed-off-by: Oliver Pinter <oliver.pinter@hardenedbsd.org>