Add an irf_len field to struct if_rxd_frag
Each frag can have a separte length, and we don't know what it will be
especially since the driver doesn't (currently) know the size of the buffers.

This makes if_rxd_info.iri_len redundant, but I haven't taken the time to clean
that out, so for now, both must be correct.